requirements.txt
torch
transformers
streamlit
pillow
requests
